requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
slack-sdk==3.26.1
aiohttp==3.9.1
openai==1.97.0
httpx==0.27.0
python-pptx==0.6.23
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
pypdf==3.17.0
pdf2image==1.16.3
Pillow==10.1.0
reportlab==4.0.7